Specextract fitting accounts for Si line

+++ code/spec_plot2pdf.sh
@@ -2,6 +2,7 @@
 
 # Short script to batch convert ps2pdf and remove intermediate .ps files
 # Aaron Tran, 2014 June 17
+# Modified 2014 June 23 (gracefully exit on finding 0 files)
 
 echo "Enter file stem (e.g., stem for ./plots/plt_src1_grp.ps is /plots/plt):"
 read stem
@@ -9,12 +10,17 @@ read stem
 nfiles=`ls -l ${stem}_src*_grp.ps | wc -l`
 echo "${nfiles} .ps files identified"
 
-for i in `seq 1 ${nfiles}`;
-do
-    infile=${stem}_src${i}_grp.ps
-    outfile=${stem}_src${i}_grp.pdf
-    echo "Converting ${infile} to ${outfile}"
-    ps2pdf $infile $outfile
-    rm $infile
-done
-echo Done!
+if [ $nfiles != 0 ]
+then
+    for i in `seq 1 ${nfiles}`;
+    do
+        infile=${stem}_src${i}_grp.ps
+        outfile=${stem}_src${i}_grp.pdf
+        echo "Converting ${infile} to ${outfile}"
+        ps2pdf $infile $outfile
+        rm $infile
+    done
+    echo Done!
+else
+    echo "No files to convert, exiting"
+fi
